test: tighten eeBUS config boundary
